Extract dashboard mock helpers
