+++ client/web/src/integration/insights/insight/dashboard-cards.test.ts
@@ -1,15 +1,8 @@
 import assert from 'assert'
 
 import { createDriverForTest, Driver } from '@sourcegraph/shared/src/testing/driver'
-import { testUserID } from '@sourcegraph/shared/src/testing/integration/graphQlResults'
 import { afterEachSaveScreenshotIfFailed } from '@sourcegraph/shared/src/testing/screenshotReporter'
 
-import {
-    GetDashboardInsightsResult,
-    GetInsightViewResult,
-    InsightsDashboardNode,
-    InsightViewNode,
-} from '../../../graphql-operations'
 import { createWebIntegrationTestContext, WebIntegrationTestContext } from '../../context'
 import {
     CAPTURE_GROUP_INSIGHT,
@@ -21,7 +14,7 @@ import {
     LANG_STATS_INSIGHT,
     SEARCH_BASED_INSIGHT,
 } from '../fixtures/dashboards'
-import { OverrideGraphQLExtensionsProps, overrideInsightsGraphQLApi } from '../utils/override-insights-graphql-api'
+import { mockDashboardWithInsights } from '../utils/mock-helpers'
 
 describe('Code insights [Dashboard card]', () => {
     let driver: Driver
@@ -48,6 +41,10 @@ describe('Code insights [Dashboard card]', () => {
             testContext,
             dashboardId: 'DASHBOARD_WITH_LANG_INSIGHT',
             insightMock: LANG_STATS_INSIGHT,
+            overrides: {
+                // Mock lang stats insight content
+                LangStatsInsightContent: () => LANG_STAT_INSIGHT_CONTENT,
+            },
         })
 
         await driver.page.goto(driver.sourcegraphBaseUrl + '/insights/dashboards/DASHBOARD_WITH_LANG_INSIGHT')
@@ -157,113 +154,6 @@ describe('Code insights [Dashboard card]', () => {
         await checkFilterMenu(driver, true)
     })
 })
-
-interface DashboardMockOptions {
-    id?: string
-    title?: string
-}
-
-/**
- * Creates dashboard mock entity in order to mock insight dashboard configurations.
- * It's used for easier mocking dashboards list gql handler see InsightsDashboards entry point.
- */
-function createDashboard(options: DashboardMockOptions): InsightsDashboardNode {
-    const { id = '001_dashboard', title = 'Default dashboard' } = options
-
-    return {
-        __typename: 'InsightsDashboard',
-        id,
-        title,
-        grants: {
-            __typename: 'InsightsPermissionGrants',
-            users: [testUserID],
-            organizations: [],
-            global: false,
-        },
-    }
-}
-
-interface DashboardViewMockOptions {
-    id?: string
-    insightsMocks?: InsightViewNode[]
-}
-
-/**
- * Creates mocks for the dashboard view entity, you may ask what difference between createDashboard
- * and this mock helper. Dashboard view mock contains insight configurations, dashboard mocks contain
- * only insight ids. (we should mock dashboard view only when this dashboard is opened on the page)
- */
-function createDashboardViewMock(options: DashboardViewMockOptions): GetDashboardInsightsResult {
-    const { id = '001_dashboard', insightsMocks = [] } = options
-
-    return {
-        insightsDashboards: {
-            __typename: 'InsightsDashboardConnection',
-            nodes: [
-                {
-                    __typename: 'InsightsDashboard',
-                    id,
-                    views: {
-                        __typename: 'InsightViewConnection',
-                        nodes: insightsMocks.map(mock => ({ ...mock, __typename: 'InsightView' })),
-                    },
-                },
-            ],
-        },
-    }
-}
-
-interface MakeOverridesOptions {
-    testContext: WebIntegrationTestContext
-    dashboardId: string
-    insightMock: InsightViewNode
-    insightViewMock?: GetInsightViewResult
-    overrides?: OverrideGraphQLExtensionsProps['overrides']
-}
-
-/**
- * Helper function to remove some of the boiler plate when overriding gql calls.
- */
-function mockDashboardWithInsights({
-    testContext,
-    dashboardId,
-    insightMock,
-    insightViewMock,
-    overrides,
-}: MakeOverridesOptions) {
-    const defaultOverrides: OverrideGraphQLExtensionsProps['overrides'] = {
-        // Mock list of possible code insights dashboards on the dashboard page
-        InsightsDashboards: () => ({
-            currentUser: {
-                __typename: 'User',
-                id: testUserID,
-                organizations: { nodes: [] },
-            },
-            insightsDashboards: {
-                __typename: 'InsightsDashboardConnection',
-                nodes: [createDashboard({ id: dashboardId })],
-            },
-        }),
-        // Mock dashboard configuration (dashboard content) with one capture group insight configuration
-        GetDashboardInsights: () =>
-            createDashboardViewMock({
-                id: dashboardId,
-                insightsMocks: [insightMock],
-            }),
-
-        // Mock lang stats insight content
-        LangStatsInsightContent: () => LANG_STAT_INSIGHT_CONTENT,
-    }
-
-    if (insightViewMock) {
-        defaultOverrides.GetInsightView = () => insightViewMock
-    }
-
-    overrideInsightsGraphQLApi({
-        testContext,
-        overrides: { ...defaultOverrides, ...overrides },
-    })
-}
 
 /**
  * Asserts the options menu renders the correct options
